pyproject.toml
[project]
name = "ai-powered-cloud-cost-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "langchain-community>=0.4.1",
    "langchain[huggingface]>=1.2.0",
    "msoffcrypto-tool>=5.4.2",
    "networkx>=3.6.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "unstructured>=0.18.21",
]
